infoportal_app/requirements.txt
alembic==1.8.1
asyncpg==0.25.0
psycopg2==2.9.3
click==8.1.3
databases==0.6.0
fastapi==0.85.1
fastapi-jwt-auth==0.5.0
httptools==0.4.0
loguru==0.6.0
ormar==0.11.3
pydantic==1.9.1
PyJWT==1.7.1
python-dotenv==0.20.0
PyYAML==6.0
SQLAlchemy==1.4.38
uvicorn==0.18.3
uvloop==0.16.0
fastapi-utils==0.2.1
fastapi_mail==1.2.0
